feat: Project RPG
능력치 추가 및 관련 시스템 정리

+++ 02_CSharpAlgorithm/PersonalProject/ProjectRPG/Interfaces/ITargetable.cs
@@ -11,6 +11,6 @@
         /// </summary>
         /// <param name="damage">가한 데미지</param>
         /// <returns>피격 성공 여부</returns>
-        bool Hit(float damage);
+        bool Hit(int damage);
     }
 }
